Introduce logging mechanism. Fix some bugs

+++ src/main/java/com/master/milano/repository/ItemInterestRepository.java
@@ -9,9 +9,10 @@ import java.util.Collection;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface ItemInterestRepository extends CrudRepository<ItemInterest, Integer> {
+public interface ItemInterestRepository extends CrudRepository<ItemInterest, Long> {
 
     Optional<ItemInterest> findByUserIdAndItem(String userId, Item item);
+
 
     @Query(
             value = "SELECT ii.* FROM item_interest ii join item i on ii.item_id = i.id  WHERE i.public_id=?1",
